delete pag (problemas!)

+++ src/app/websites/websites.component.ts
@@ -78,8 +78,14 @@ export class WebsitesComponent {
   }
 
   delete(website: Website): void {
-    this.websites = this.websites.filter(h => h !== website);
-    this.websiteService.deleteWebsite(website._id).subscribe();
+    if (website.urls && website.urls.length > 0) {
+      if (confirm('Este website tem páginas associadas. Tem certeza de que deseja excluí-lo?')) {
+        this.websites = this.websites.filter(h => h !== website);
+        this.websiteService.deleteWebsite(website._id).subscribe();
+      }
+    } else {
+      this.websites = this.websites.filter(h => h !== website);
+      this.websiteService.deleteWebsite(website._id).subscribe();
+    }
   }
-
 }
